% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSPDF.R
\name{createSPDF}
\alias{createSPDF}
\title{Internal helper function}
\usage{
createSPDF()
}
\value{
An empty SpatialPolygonsDataFrame object matching the expected schema of the estKIN function.
}
\description{
Create empty SpatialPolygonsDataFrame with estObj schema
}
\author{
Shannon E. Albeke, Wyoming Geographic Information Science Center, University of Wyoming
}
