% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColors.R
\name{getColors}
\alias{getColors}
\title{Create a list of colors for plotKIN function}
\usage{
getColors(groups, levels)
}
\arguments{
\item{groups}{The number of groups within grouping variable (i.e. species)}

\item{levels}{The number of confidence intervals provided by the user}
}
\value{
A character vector of RGB colors
}
\description{
The list of colors were obtained from Colorbrewer2.org using single hue.
This is run within the function plotKIN()
}
\author{
Shannon E. Albeke, Wyoming Geographic Information Science Center, University of Wyoming
}

