% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKernelThreshold.R
\name{getKernelThreshold}
\alias{getKernelThreshold}
\title{Miscellaneous functions to complete kernel 2D estimates: Get contour threshold values}
\usage{
getKernelThreshold(x, levels = c(50, 75, 95))
}
\arguments{
\item{x}{Numeric vector of probabilities from a kernel density estimate}

\item{levels}{Numeric vector of desired percent levels (e.g. c(10, 50, 90). Should not be less than 1 or greater than 99)}
}
\value{
A list of threshold values for each percent.
}
\description{
Obtains the quantile threshold levels for a vector of probabilities from a
kernel density estimate.
}
\author{
Shannon E. Albeke, Wyoming Geographic Information Science Center, University of Wyoming
}

