% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHole.R
\name{makeHole}
\alias{makeHole}
\title{Miscellaneous functions to complete kernel 2D estimates: make polygon hole when inner circle}
\usage{
makeHole(poly, outer, hole)
}
\arguments{
\item{poly}{The multipart polygon}

\item{outer}{The outer ring of the polygon}

\item{hole}{The inner ring of the polygon}
}
\value{
A Polygons object
}
\description{
For multipart polygons, set the inner circle to be 'hole=TRUE' after obtaining contour lines from kernel estimate.
The hole is determined using the gContainsProper() function from rgeos package. This is run within the
2D-Kernel estimator function estKIN()
}
\author{
Shannon E. Albeke, Wyoming Geographic Information Science Center, University of Wyoming
}
