% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-rKIN.R
\docType{package}
\name{rKIN}
\alias{rKIN}
\alias{rKIN-package}
\title{rKIN: A package for computating isotopic niche space}
\description{
The rKIN This package applies methods used to estimate animal homerange, but
instead of geospatial coordinates, we use isotopic coordinates. The estimation
methods include: 1) 2-dimensional bivariate normal kernel utilization density
estimator, 2) bivariate normal ellipse estimator, and 3) minimum convex polygon
estimator, all applied to stable isotope data. Additionally, functions to
determine niche area, polygon overlap between groups and levels (confidence
contours) and plotting capabilities.
}
\section{rKIN functions}{

The rKIN functions:
estKIN, estEllipse, estMCP, plot.kin, getArea, calcOverlap
}

