% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rodents}
\alias{rodents}
\title{Isotopic data from rodent blood samples.}
\format{A data frame with 530 rows and 4 variables:
\describe{
  \item{Species}{Generic species used as a grouping variable}
  \item{Habitat}{Habitat in which the individual was captured}
  \item{Ave_C}{Averaged delta 13C present within the blood sample}
  \item{Ave_N}{Averaged delta 15N present within the blood sample}
  ...
}}
\source{
\url{http://www.uwyo.edu/zoology/people/bendavid.html}
}
\usage{
rodents
}
\description{
A dataset containing the individual Species, Habitat sampled,
 and Percent delta C and N.
}
\keyword{datasets}
